requirements.txt
fastapi==0.65.1
uvicorn==0.13.4
python-multipart==0.0.5
Pillow==8.2.0
numpy==1.20.3
